Spoitfy command & fix Say command

+++ src/Jacky/Commands/SayCommand.php
@@ -22,12 +22,15 @@ class SayCommand extends Command implements CommandInterface
 
     public function execute(Message $message, $params = [])
     {
+        $delete = false;
         $whatToSay = CommandHelper::removeMessageMentions($this->getContent($message->content), $message->mentions);
         foreach ($message->mentions as $mention) {
             $message->channel->sendMessage(sprintf("%s %s (de la part de %s)", $mention, $whatToSay, $message->author));
+            $delete = true;
         }
 
-        $message->channel->messages->delete($message);
+        if($delete)
+            $message->channel->messages->delete($message);
     }
 
     public function getName()
